# Mechanic_App/static/swagger.yaml
swagger: '2.0'
info:
  title: "Mechanics_API"
  description: "This is an API to manage a Mechanic shop database"
  version: "1.0.0"
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
paths:
...
  /customers/login:
    post:
      tags:
        - Customers
      summary: "Login endpoint, generates a token"
      description: "Validates customer credentials and returns a token."
      parameters:
        - in: "body"
          name: "body"
          required: true
          description: "Login credentials"
          schema:
            $ref: "#/definitions/CustomerLoginCredentials"
      responses:
        200:
          description: "Login successful"
          schema:
            $ref: "#/definitions/CustomerLoginResponse"
  /customers: #endpoint
    post: #This is a post request
      tags:
        - Customers
      summary: "This endpoint create a new user"
      description: "This endpoint create a new customer in our Mechanic system, the emails must be unique(you can't use one email twice)"
      parameters:
        - in: "body"
          name: "body"
          description: "Information required to create a new customer"
          schema: 
            $ref: "#/definitions/CreateCustomerPayload"
      responses:
        201:
          description: "Successfull Custumer Creation"
          schema:
            $ref: "#/definitions/CreateCustomerResponse"
          examples:
            application/json:
              id: 1
              name: Jane Doe
              phone: 000-000-0000
              email: janed@email.com
              password: pass123

    get:
      tags:
        - Customers
      summary: "Returns all Customers"
      description: "Endpoint to retrieve the list of all customers in the Mechanic database"
      responses:
        200:
          description: "Retrieved Customers Successfully"
          schema:
            $ref: "#/definitions/RetrieveAllCustomers"

    put:
      tags:
        - Customers
      summary: "Updates user information"
      description: "Endpoint to update a customer, this is a token authenticated route, and the user must be logged in to update their account."
      security:
        - bearerAuth: []
      parameters:
        - in: "body"
          name: "body"
          description: "Information required to update a customer"
          required: true
          schema:
            $ref: "#/definitions/UpdateCustomerPayload"
      responses:
        200:
          description: "Successfully updated a Customer"
          schema:
            $ref: "#/definitions/UpdateCustomerResponse"
          examples:
            application/json:
              id: 1
              name: Jane Doe
              phone: 000-000-0000
              email: janed@email.com
              password: pass123

    delete:
      tags:
        - Customers
      summary: "Endpoint to delete a customer"
      description: "Endpoint to delete a customer. Requires token authentication, customer must be logged in to delete their account"
      security:
        - bearerAuth: []
      responses:
        200:
          description: "Successfully deleted a customer"
      schema:
        $ref: "#/definitions/DeleteCustomerResponse"
      examples:
        application/json:
          message: "Successfully deleted user <customer_id>"
  /customers/{customer_id}:
    get:
      tags:
        - Customers
      summary: "Retrieves a single customer"
      description: "This endpoint retrieves a single customer from the Mechanic database"
      parameters:
        - name: customer_id
          in: path
          required: true
          type: integer
          description: "The ID of the customer that you want retrieve"
      responses:
        200:
          description: "Retrieve customer Successfully"
          schema:
            $ref: "#/definitions/RetrieveACustomer"
  /mechanics/login:
    post:
      tags:
        - Mechanics
      summary: "Endpoint to login a mechanic and it generates token"
      description: "Validates customer credentials and returns a token."
      parameters:
        - in: "body"
          name: "body"
          required: true
          description: "Mechanic login credentials"
          schema:
            $ref: "#/definitions/MechanicLoginCredentials"
      responses:
        200:
          description: "Login Successful"
          schema: 
            $ref: "#/definitions/MechanicLoginResponse"
  /mechanics:
    post:
      tags:
        - Mechanics
      summary: "Creates a user"
      description: "This endpoint creates a mechanic employee. Make sure your email is unique, you are not allow to use one email twice"
      parameters:
        - in: "body"
          name: "body"
          required: true
          description: "Information required to create a mechanic employee"
          schema:
            $ref: "#/definitions/CreateMechanicPayload"
      responses:
        201:
          description: "Successful mechanic creation"
          schema:
            $ref: "#/definitions/CreateMechanicPayload"
          examples:
            application/json:
              id: 1
              name: John Doe
              email: johnd@email.com
              phone: 000-000-0000
              salary: 100.00
              password: password123

    get:
      tags:
        - Mechanics
      summary: "Retrieves all the Mechanics"
      description: "This endpoint retrieves all the mechanics that are currently in the database"
      responses: 
        200:
          description: "Retrieve mechanics successfully"
          schema:
            $ref: "#/definitions/RetrieveAllMechanics"

    put:
      tags:
        - Mechanics
      summary: "Updates a mechanic information"
      description: "This endpoint is used to update a mechanic. The user must be login to update their account because this is a token authenticated route"
      security: 
        - bearerAuth: []
      parameters:
        - in: "body"
          name: "body"
          description: "Information required to update a mechanic"
          required: true
          schema:
            $ref: "#/definitions/UpdateMechanicPayload"
      responses:
        200:
          description: "Successfully updated mechanic information"
          schema:
            $ref: "#/definitions/UpdateMechanicResponse"
          examples:
            application/json:
              id: 1
              name: John Doe
              phone: 000-000-0000
              email: johnd@gmail.com
              password: password123

    delete:
      tags:
        - Mechanics
      summary: "Endpoint deletes the mechanic"
      description: "This endpoint is used to delete a mechanic. It requires token authentication, the mechanic must be logged in"
      security:
        - bearerAuth: []
      responses:
        200:
          description: "Successfully deleted mechanic"
          schema:
            $ref: "#/definitions/DeleteMechanicResponse"
          example:
            application/json:
              message: "successfully deleted mechanic <mechanic_id>."
...
definitions:
...
  CustomerLoginCredentials:
    type: "object"
    properties:
      email:
        type: "string"
      password:
        type: "string"
    required:
      - email
      - password
  CustomerLoginResponse:
    type: "object"
    properties:
      token:
        type: "string"
      message:
        type: "string"
      status:
        type: "string"
  CreateCustomerPayload:
    type: "object"
    properties:
      name:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      password:
        type: "string"

    required:
      - name
      - email
      - phone
      - password
  CreateCustomerResponse:
    type: "object"
    properties:
      id: 
        type: "integer"
      name:
        type: "string"
      phone:
        type: "string"
      password:
        type: "string"
  RetrieveAllCustomers:
    type: "array" #use array because we are getting a list of customers
    items:
      type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      phone:
        type: "string"
      password:
        type: "string"
  RetrieveACustomer:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      phone:
        type: "string"
      password:
        type: "string"
  UpdateCustomerPayload:
    type: "object"
    properties:
      name:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      password:
        type: "string"
    
    required:
      - name
      - email
      - phone
      - password
  UpdateCustomerResponse:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      password:
        type: "string"
  DeleteCustomerResponse:
    type: "object"
    properties:
      message:
        type: "string"

#Mechanic Definition
  MechanicLoginCredentials:
    type: "object"
    properties:
      email:
        type: "string"
      password:
        type: "string"

    required:
      - email
      - password
  MechanicLoginResponse:
    type: "object"
    properties:
      token:
        type: "string"
      message:
        type: "string"
      status:
        type: "string"
  CreateMechanicPayload:
    type: "object"
    properties:
      name:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      salary:
        type: "string"
      pasword:
        type: "string"
...
  RetrieveAllMechanics:
    type: "array"
    items:
      type: "object"
      properties:
        id:
          type: "integer"
        name:
          type: "string"
        email:
          type: "string"
        phone:
          type: "string"
        salary:
          type: "string"
        password:
          type: "string"
  UpdateMechanicPayload:
    types: "object"
    properties:
      name:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      salary:
        type: "string"
      password:
        type: "string"
  UpdateMechanicResponse:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      salary:
        type: "string"
      password:
        type: "string"
  DeleteMechanicResponse:
    type: "object"
    properties:
      message: #don't add the rest of the properties in the delete response
        type: "string"
...